Move the Sierra HTTP client into the shared lib
